requirements.txt
sqlalchemy
python-dotenv
fastapi
pydantic
sentry-sdk[fastapi]
# sentry-sdk
uvicorn
